OVC5 Carrier Firmware (#54)
* WIP Vivado project for the carrier board


* Swap 0/1 inputs


* Update petalinux config


* Update device tree


* dhcpd starts correctly on boot, USB gadget is hot-pluggable


* uio device mem location output


* Triggers are now driver via GPIO as TRIG0 cannot be configured as an input


* GPIO Driver initial code


* Adjustments after testing


* Enable cameras via gpio driver


* Change error handling for read/write


* Add escapement from indefinite read


* Updated for OVC5 custom carrier board


* Format command and fix issues


* Sampling triggers


* Unexport gpio when finished with a pin


* Fixed LED, pin cleanup, and added sleep to give time for cameras to initialize


* Trigger before reading out next frames.


* Format fix


* Camera device number


* Re-enable flag


* Line counter per camera (debug attached to cam 2 and 5)


* Single line counter selected by GPIO


* Adjusting to 4k resolution and correcting camera indexes


* Fixed connection issue (cameras now working)


* Configs for cam 2 and 5


* Fixed issue with bus connections on all rx blocks


* Fix interfaces misconfiguration


* Make configs grouped by camera instead of index in array (#2)

* Configs are divided by camera


* Add the ability to run device driver from makefile


* Ignore generated psu c files


* PWM Output drives triggers. Correct cam3 rx.


* Adding usb1 ethernet gadget scripts

+++ ovc5/software/libovc/src/ovc.cpp
@@ -1,5 +1,6 @@
 #include "libovc/ovc.hpp"
 
+#include <iomanip>
 #include <thread>
 
 #include "libovc/ethernet_packetdef.hpp"
@@ -14,7 +15,7 @@ static cv::Mat unpackTo16(const cv::Mat &frame)
   uint16_t *out16 = (uint16_t *)ret.data;
 
   size_t n = frame.cols * frame.rows;
-  unsigned full, half;
+  uint16_t full, half;
 
   for (n /= 2; n--;)
   {
@@ -44,7 +45,6 @@ OVC::~OVC()
 std::array<OVCImage, Server::NUM_IMAGERS> OVC::getFrames()
 {
   frames_ = server_.getFrames();
-
   for (size_t i = 0; i < frames_.size(); i++)
   {
     cv::Mat shifted = unpackTo16(frames_[i].image);
